django/markdown/requirements.txt
asgiref==3.3.1
bleach==3.2.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-markdownify==0.8.1
django-pygments==0.4
Markdown==3.3.3
packaging==20.8
Pygments==2.7.3
pyparsing==2.4.7
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
